+++ server/CHANGELOG.md
@@ -3,3 +3,5 @@
 ## 0.0.1
 
 - Initial Version
+- Dotenv validation
+- Bare minimum express with CORS is setup